move min labware definitions to conftest.
