+++ api/tests/opentrons/protocol_api/test_accessor_fn.py
@@ -1,156 +1,4 @@
-import pytest
-from opentrons.protocol_api import labware
-from opentrons.types import Point, Location
-
-
-@pytest.fixture
-def minimal_labware_def():
-    return {
-        "metadata": {
-            "displayName": "minimal labware"
-        },
-        "cornerOffsetFromSlot": {
-            "x": 10,
-            "y": 10,
-            "z": 5
-        },
-        "parameters": {
-            "isTiprack": False,
-            "loadName": "minimal_labware_def"
-        },
-        "ordering": [["A1"], ["A2"]],
-        "wells": {
-            "A1": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 0,
-              "y": 0,
-              "z": 0,
-              "shape": "circular"
-            },
-            "A2": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 10,
-              "y": 0,
-              "z": 0,
-              "shape": "circular"
-            }
-        },
-        "dimensions": {
-            "xDimension": 1.0,
-            "yDimension": 2.0,
-            "zDimension": 3.0
-        }
-    }
-
-
-@pytest.fixture
-def minimal_labware_def2():
-    return {
-        "metadata": {
-            "displayName": "other test labware"
-        },
-        "cornerOffsetFromSlot": {
-                "x": 10,
-                "y": 10,
-                "z": 5
-        },
-        "parameters": {
-            "isTiprack": False,
-            "loadName": "minimal_labware_def"
-        },
-        "ordering": [["A1", "B1", "C1"], ["A2", "B2", "C2"]],
-        "wells": {
-            "A1": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 0,
-              "y": 18,
-              "z": 0,
-              "shape": "circular"
-            },
-            "B1": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 0,
-              "y": 9,
-              "z": 0,
-              "shape": "circular"
-            },
-            "C1": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 0,
-              "y": 0,
-              "z": 0,
-              "shape": "circular"
-            },
-            "A2": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 9,
-              "y": 18,
-              "z": 0,
-              "shape": "circular"
-            },
-            "B2": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 9,
-              "y": 9,
-              "z": 0,
-              "shape": "circular"
-            },
-            "C2": {
-              "depth": 40,
-              "totalLiquidVolume": 100,
-              "diameter": 30,
-              "x": 9,
-              "y": 0,
-              "z": 0,
-              "shape": "circular"
-            }
-        },
-        "dimensions": {
-            "xDimension": 1.0,
-            "yDimension": 2.0,
-            "zDimension": 3.0
-        }
-    }
-
-
-@pytest.fixture
-def min_lw_impl(minimal_labware_def) -> labware.LabwareImplementation:
-    return labware.LabwareImplementation(
-            definition=minimal_labware_def,
-            parent=Location(Point(0, 0, 0), 'deck')
-    )
-
-
-@pytest.fixture
-def min_lw2_impl(minimal_labware_def2) -> labware.LabwareImplementation:
-    return labware.LabwareImplementation(
-        definition=minimal_labware_def2,
-        parent=Location(Point(0, 0, 0), 'deck')
-    )
-
-
-@pytest.fixture
-def min_lw(min_lw_impl) -> labware.Labware:
-    return labware.Labware(implementation=min_lw_impl)
-
-
-@pytest.fixture
-def min_lw2(min_lw2_impl) -> labware.Labware:
-    return labware.Labware(implementation=min_lw2_impl)
+from opentrons.types import Point
 
 
 def test_labware_init(min_lw_impl, minimal_labware_def):
